stop jOOQ narrowing every stored decimal to Float (#116)
jOOQ's SQLite dialect maps REAL to Float, so every stored decimal without an
exact 32-bit form was degraded on the round-trip: a 107.05 trend weight read
back 107.05000305175781. SQLite REAL is already an 8-byte double, so the
column type was never the problem — the type *mapping* was.

A single <forcedType> on REAL fixes it in one place, which then makes all 41
manual .toFloat()/.toDouble() conversions across the seven repositories
redundant. Zero manual conversions now remain in persistence: the DSL.sum()
BigDecimal hop went too, by letting jOOQ convert via fetchOne(0, Double).

Coverage is exact: 18 REAL columns generate 18 Double fields, no Float
remains, and the 15 Integer columns are untouched.

The tolerances this bug forced into the suite are tightened rather than a new
test being invented for its own sake — the asymmetry between 1e-2 on persisted
doubles and 1e-6 on computed ones *was* the bug. Reverting only the fix now
reddens 7 tests, including "a fresh Goal reads exactly 0%", which is ADR 0016's
own wording (line 28) that the narrowing had quietly falsified.

No schema migration, no API change (the OpenAPI spec regenerates byte-identical).
Rows written before this fix keep their narrowed value and are not backfilled;
reading them is unaffected, since float->double->float is identity and every
removed conversion was on the write side.

+++ backend/src/main/kotlin/com/tucker/persistence/ProfileRepository.kt
@@ -20,7 +20,7 @@ class ProfileRepository(private val dsl: DSLContext) {
             dsl.update(PROFILE)
                 .set(PROFILE.SEX, profile.sex.name)
                 .set(PROFILE.BIRTH_DATE, profile.birthDate.toString())
-                .set(PROFILE.HEIGHT_CM, profile.heightCm.toFloat())
+                .set(PROFILE.HEIGHT_CM, profile.heightCm)
                 .set(PROFILE.TIMEZONE, profile.timezone)
                 .set(PROFILE.REMINDER_HOUR, profile.reminderHour)
                 .set(PROFILE.REMINDERS_ENABLED, profile.remindersEnabled.toFlag())
@@ -31,7 +31,7 @@ class ProfileRepository(private val dsl: DSLContext) {
                 .set(PROFILE.ID, SINGLETON_ID)
                 .set(PROFILE.SEX, profile.sex.name)
                 .set(PROFILE.BIRTH_DATE, profile.birthDate.toString())
-                .set(PROFILE.HEIGHT_CM, profile.heightCm.toFloat())
+                .set(PROFILE.HEIGHT_CM, profile.heightCm)
                 .set(PROFILE.TIMEZONE, profile.timezone)
                 .set(PROFILE.REMINDER_HOUR, profile.reminderHour)
                 .set(PROFILE.REMINDERS_ENABLED, profile.remindersEnabled.toFlag())
@@ -42,7 +42,7 @@ class ProfileRepository(private val dsl: DSLContext) {
     private fun ProfileRecord.toProfile(): Profile = Profile(
         sex = Sex.valueOf(sex),
         birthDate = LocalDate.parse(birthDate),
-        heightCm = heightCm.toDouble(),
+        heightCm = heightCm,
         timezone = timezone,
         reminderHour = reminderHour,
         remindersEnabled = remindersEnabled != 0,
